[fix]: last commit is change the permissions - this change the data type ssize_t to int in writep

+++ 0x15-file_io/1-create_file.c
@@ -11,7 +11,7 @@ int create_file(const char *filename, char *text_content)
 {
 	int fd;
 	int len;
-	ssize_t writep;
+	int writep;
 
 	if (filename == NULL)
 		return (-1);
